requirements.txt
starlette>=0.27
uvicorn>=0.24
httpx>=0.25
pydantic-settings>=2.0
python-jose[cryptography]>=3.3
passlib>=1.7
slowapi>=0.1.9
